Completed Date and Time setting
Date and Time can now be set from the app.  It's located under the Memory command.  There are several options to choose.  You can query the date and time from the controller, set the current time to the controller based on your device's current date and time or you can specify a custom date and time.