Fixed bug with swipe and 2 or more expansion relays
When the swipe would destroy pages and recreate them, our custom scrollview that we created was causing a problem of the page already existing when trying to readd the page.  Changed all the pager widgets (Controller and Relays) to be scrollview widgets to solve the problem.
Added pager support to landscape view
Changed debugging text for menu clicks