Scroll to Status screen when entereing preferences
This is to ensure that the screen being displayed is always supposed to be there and visible.  The primary reason is if somebody changes the expansion relay quantity to a quantity less than the current quanitity, the last relay page will be displayed and will "jump" back to the previous page.
This will also benefit future improvements with additional screens and does not hurt anything to start back at the main screen once exiting the Preferences page since the user most likely changed something.