Fixed parsing failure
The problem was using a byte variable to store a value greater than 127.  Android uses byte as a signed byte.  Some of the values that we are parsing into a byte are greater than 127.  This would cause the XML parser to fail when we were doing the conversion.  Switched the byte's over to short's to fix the problem.