Added new internal memory location starting point
With v0.9.9, the internal memory starting point changed from 800 to 200.  Added in the ability to control what starting point to use for the internal memory when referencing it.  The default is to use the Old / Pre-0.9.9 locations.  You can simply uncheck the setting to use the new locations.  The default was set to pre-0.9.9 to not break all the existing installations of the app.  This default selection can and should change in the future.