Fixed parsing bug
Check for status xml data vs label data failed with RON.  The check was to see if N was the last char in the tag indicating a label.  RON broke that and would process the data as a label and error on parsing.
Fixed warning about data type mismatches with shorts and bytes.