Fixed page switching based on installed modules
Store the pages in an array.  When we make changes to the installed modules, we then update the array of modules installed.  The order of the array is the order in which the pages are displayed.  When the pager adapter calls instantiateItem, there is no calculations needed since the position already correlates with the proper order.
Closes #41