Updated database storage of values
Changed AP, DP & Dimming expansion channels to store integers (shorts) instead of text.  The values stored now contain the values and not the % sign as well.  The application displaying the data is responsible for showing the proper % signs. Also, the controller class no longer returns the units with salinity and orp.  This will not affect anything but the status application.