Auto update modules
Automatically updates the modules enabled in the app based on the settings from the controller.  This will override any current settings in your app.  This is enabled by default.  You can disable it if you choose but it's best left enabled so you don't have to reconfigure your app whenever you change things with your controller.
Closes #10