Auto Updating Changes
Switched to update service and using alarm manager to signal the updates.
Added ability to choose what profile to perform the auto updates on: Always, Only on Away, Only on Home.  Only on Away profile is the default.
Auto update only runs on Only on Away or Only on Home when the proper profile is selected (Home or Away).
Stop and start the auto update service appropriately when you change profiles, devices, hosts, ports, intervals.
Closes #56
Closes #55