Updated how service is handled
Main service that the application uses is not started when the device boots. Only the auto update service is started. When the application is paused, the service is also paused except when launching an activity inside the app (Memory, Commands, Settings). The service is started when the application resumes. There are 2 exceptions: 1. if the previously mentioned activities are paused, then the service will stay running unless you quit it or return back to the main app and pause it. The service does not get stopped when the Memory, Commands or Settings activities are paused and does not get restarted from those activies. To restart the service, you must simply return to the main Status activity and all will be well. 2. If you launch the Forum or RA Website from the preferences, the service is not stopped.
Removed some extra debugging statements.