Add try/catch around relay tag processing
This will be helpful when a new tag is added that begins with R and is not currently being handled. Prior handling would generate a NumberFormatException error when trying to parse an invalid integer thus generating a SAX error and the XML parsing would fail. This will simply log the error and continue parsing the XML file.