Removed ControllerService
ControllerService was an intermediary service that just passed on commands to the UpdateService. Instead of continuously checking if the ControllerService was running and send messages to it, remove ControllerService and have the activities start the UpdateService whenever a task needs to be executed. This eliminates the need to check if the ControllerService is running and start/stop it as needed. The UpdateService handled all the communication processes and runs its tasks and then stops, so it handles itself automatically.