Added displaying of ATO status on main page
ATO status is displayed as ON / OFF depending on the status of the port. A value of 1 is equal to ON, ACTIVE, or the GREEN circle (on the portal). A value of 0 is equal to OFF, INACTIVE, or the RED circle (on the portal). Currently, on ON/OFF is displayed here but will be improved in the future to be the actual circle.
ATO visibility is toggled through the controller page visibility.
ATO labels are updated through the controller page labels. Labels are not currently downloaded from the Portal, so you must edit them manually.
Closes #25
Fixed PH Label title from displaying incorrectly on the Labels page.