Finished switching to ContentProvider
Insert and Delete works. History page gets updated as more parameters are logged. Cleaned up the code and removed unecessary functions. Inserts are now handled from within ControllerTask instead of calling a function inside the application and having to set and read all the data from a ContentValues. This means improved memory usage and performance should be improved even if it is slightly.