Updated error notification message
Instead of broadcasting the error message in an Intent, have the notifyuser function grab the error messge to display it. The main status activity no longer will show the complete error message. Instead, it will only show Error and the full message will be displayed in the Notification Bar.

This is in the first phase of the improved notification handling and displaying scheme of the app.