Changed how we read from the xml stream
Initially we were reading the entire xml file into memory by allocating a new byte array for the amount of data available on the inputstream. Then we were passing that string into the xml parser to be parsed.
A better way to handle this is to let the parser do the reading directly from the input stream of the connection. This cuts out the extra memory requirements and usages.

This commit only bypasses the reading of the xml file into memory. Further code cleanup and status updates are needed. Code is working but needs improved.