Updated page data on swipes
Added static functions for final classes that have references to the packages. Allow them to add in .debug for the debug builds to match the rest of the packages.
Added in PageInterface for the fragments.
On page scroll calls the fragment interface to refresh the data.
Removed the category for AutoUpdate preference screen.