Moved Get Version to Commands Page
Get Version was at the bottom of the commands activity page and now it is moved to the Commands Page in the main activity. Pressing the button queries the controller for the Version and the button's label is updated with the version of the controller's software.
Removed the unused files for the tab activity
closes #33