cleaned up override code
moved check to see if pwm port is overriden to a static function inside the controller class. the code was saving the override value as the actual value and then saving the override value in the database. changed the code around to be storing the actual value and the override value in their respective columns (like they should be). then when we display the values, have the new static functions determine what value to be displayed and indicate if the channel is overridden.

currently, only the ap and dp values work like this. the other pwm values will only display their original values.