Added in displaying pwm override popup
Displays popup that shows the current value. It allows you to change the value with a slider and it updates the percentage in the window. You can then either Set the new value, Clear the Override or cancel the dialog. The dialog itself does not indicate if the override is currently set, you must look at the main window to see if it is set.
Only DP and AP handle this request