Handle override response from the controller
Response comes back like the memory commands. It responds <P#>MSG</P#>
Where # is the channel that was altered
and MSG is the response either OK or ERR

When the response is received, the app says to refresh the screen to update.
Consider automatically refreshing the screen for the user.