Improved message for override popup
If the channel allows or has a custom name, display that name with the default channel name in parenthesis. If the channel does not allow for a custom name (radion or ai modules), just display the module name plus the channel name plus channel. ex. "Radion Blue Channel" or "Aqua Illuminations White Channel"