Added in basic webserver authentication
Added the ability to communicate with the controller running with basic webserver authentication. This functionality will only be available for version 1.1.1 of the libraries.
In the settings, you can input your username and password to be used for your controller. If either one is empty, the authentication is not used.
This functionality needs to be tested to ensure it works properly.
closes #39