Removal of ViewPagerIndicator library dependency
Removed the reference and requirement of ViewPagerIndicator library. The functionality was added into the support libraries. Also, maintaining a working build of the library across mulitple development devices was trickier.

This work was done by phrusher and sent as a pull request. It was manually added due to the changes made to the code base.