Renamed app module from AndroidStatus to app
The initial conversion to Android Studio created the project AndroidStatus and a module called AndroidStatus. The "default" naming convention for most android apps is to have the application name be the project and the module name be app. The module name was renamed to follow along with this format.
Also, the iml file was updated and the paths were simplified and cleaned up in this process.