Fixed Navigation Drawer
The navigation drawer was not functioning properly with a logo being present in the drawer. that shifted the index of all the items down 1. that was a problem displaying items appropriately. Also, the highlighting of the correct item in the list was not working properly because of the backstack. Implemented a backstack listener and handle the highlighting of the navigation item properly based on the actual page being displayed.