Implemented label downloading from portal
Added the download label functionality in the preferences screen.
Fixed error with downloading while authentication was enabled. The labels are downloaded from the Portal and that does not utilize the controller authentication. A check was needed to see if we were getting the labels from the Portal. If we are, we disable the authentication because an entirely different URL is used.