Added DC Pump page initially
Added displaying the DC Pump page from the settings OR having it enabled based on the module flags from the controller.
Added the long press capability to the values but they display the vortech dialog box (that needs to be fixed to be more open).
The display for the dc pump page needs improved. All fields displayed are just the numeric values except the speed adds a percent sign after it.
The threshold does not display a long press dialog box even though it responds to a long press event.