Updated app logo
Changed the application logo to the white reef angel logo that matches the large logo on the Play store. This was requested because Roberto's universal app uses the same logo as this app and it makes it hard to distinguish on a device.
Hide the app icon on the actionbar by replacing it with a transparent icon. This adds more space to the actionbar and begins the transition to using the new material style of no icon on the actionbar.