Updated dc pump information
Renamed the Vortech dialog layouts to mds to be shared between the vortex and dc pump dialog boxes.
updated the strings to be more generic for the mds layouts instead of vortech specific.
added a dc pump dialog that displays the ability to change the values on the fly
fixed dc pump not displaying if it is selected from the settings page. it was not being counted in the total modules installed.
added the call to display the dc pump dialog on long press.
fixed displaying the textual version of the dc pump mode. it was just displaying the numeric value.
added the dc pump mode labels to the arrays.