Fixed button enablement with device
To bring it back inline with the main app, disable the buttons when communicating with the Portal. 
Also changed the order when refreshing the button enablement is called.
Removed duplicate calls to refresh enablement. We were looping over all 8 relay buttons each time we set a port label. That was making 8 calls to a function that iterates over all 8 ports when 1 call is sufficient when we are finished updating the labels.