Updated fragments based on device
Update the Read & Write buttons in the Memory Fragment based on the type of device. If it is the portal, we do not give the user the ability to click the Read or Write buttons.
Update the buttons on the Date / Time fragment based on the type of device. If you are communicating with the portal, then you cannot change the date and time on the device because it does not exist and you are not communicating with it.
Displays a message at the bottom of the 2 pages saying that they are disabled due to portal.