Fixed Else mode crash on long press
The long press on the DC Pump Mode was causing the app to crash. The problem was the index being sent to the dialog box was the actual value for the current mode. The Else mode has a value which is greater than the index of the spinner list thus causing the crash. We need to subtract off the jump to the upper numbers like we do in a couple other places.
closes_maxima #64