Added better preference support
Converting preferences to ActionBarActivity using PreferenceFragments. Added the ability to have a "headers" page for the preferences and then load the subsequent screens. It does not handle the large screens properly in the master/detail flow. That would need to be added in later.
Increased the build version back to 15 to allow for the changes. This may change in the future but currently the way the preferences are setup, anything less than 15 api is just too complicated to make the preferences work properly. This will be revisited in the future.