Fixed ai page display bug
Royal Blue and Blue channels were being flipped. This was due to the libraries having the incorrect channel value stored in them and that value was copied into this app. The channel values were flipped and now the value is being displayed properly.
The override values were not being stored in the database correctly. They were being stored in the actual value instead of the override value field. This was causing 255% to be displayed instead of 100%. 

closes_maxima #73