Store the full RA Dynamic Host when saving the setup wizard
Initially when saving the settings from the setup wizard, the RA Dynamic DNS host was only being set to the actual value entered. So when the controller started it would have the away host set to what was entered.
An example was this: The user would enter their username as reefangel and the RA Dynamic DNS host as home. The value being saved as the away host should have been reefangel-home.myreefangel.com but instead the away host was being set to home.

closes_maxima #72