Updated padding for the pages
Removed bottom padding on standard label rows. This frees up space and allows more items on one screen.
Removed bottom padding on the Relay page to allow all the relays to fit on the screen for most devices.
Change the textAppearance from a style to an attr for a more uniform look.