Updated the way the Back button is handled
Previously the Back button would cycle through the different pages/screens to eventually return to the Main screen. This was plagued with problems and could lead to lots of extra “Back” buttons being pressed and no extra navigation. It could also lead to the wrong section being highlighted.

The new way implemented just responds with the Back button exiting the app. The user is presented with a Toast message to indicate that an additional press is required to exit the app.