added initial code for selection criteria
Added the function that parses the date range the user selected on the Settings dialog for the charts.
Selection function only creates a date range based on the user selection and logs the current date and selected range.
Further enhancements need to be made in order to get the date range working properly.

Removed the hard coded limit of 25 results for the charts.