Allow for pH memory location value override
The pH memory locations have an imposed limit. The limit is a reasonable value, however in some rare instances the user might need to exceed them. In order to exceed the limit, the user must attempt to write a value outside the limitation range (but within the INT storage value range) and receive the warning about the invalid range/value. Then the user must attempt to write the value again within 10 seconds and the value will be allowed to be written to the controller.